perf(cli): async directiveHandler() (ESM prep)
