+++ packages/cli/src/commands/generate/directive/directive.js
@@ -1,218 +1,28 @@
-import path from 'path'
-
-import camelcase from 'camelcase'
-import execa from 'execa'
-import { Listr } from 'listr2'
-import prompts from 'prompts'
-
-import { recordTelemetryAttributes } from '@redmix/cli-helpers'
-import { getConfig } from '@redmix/project-config'
-
-import c from '../../../lib/colors.js'
 import {
-  getPaths,
-  writeFilesTask,
-  transformTSToJS,
-} from '../../../lib/index.js'
-import {
-  prepareForRollback,
-  addFunctionToRollback,
-} from '../../../lib/rollback.js'
-import { validateName } from '../helpers.js'
-import { yargsDefaults } from '../yargsCommandHelpers.js'
-import {
-  createYargsForComponentGeneration,
-  templateForComponentFile,
-} from '../yargsHandlerHelpers.js'
+  createCommand,
+  createBuilder,
+  yargsDefaults,
+} from '../yargsCommandHelpers.js'
+import { createYargsForComponentGeneration } from '../yargsHandlerHelpers.js'
 
-export const files = async ({ name, typescript = false, type, tests }) => {
-  if (tests === undefined) {
-    tests = getConfig().generate.tests
-  }
+export const command = createCommand('directive')
+export const description = 'Generate a new GraphQL directive'
+export const builder = createBuilder({ componentName: 'directive' })
 
-  if (!type) {
-    throw new Error('You must specify a directive type')
-  }
-
-  const camelName = camelcase(name)
-
-  const outputFilename = `${camelName}.${typescript ? 'ts' : 'js'}`
-
-  const directiveFile = await templateForComponentFile({
-    name,
-    extension: typescript ? '.ts' : '.js',
-    generator: 'directive',
-    templatePath: `${type}.directive.ts.template`,
-    outputPath: path.join(getPaths().api.directives, camelName, outputFilename),
-    templateVars: { camelName },
-  })
-
-  const files = [directiveFile]
-
-  if (tests) {
-    const testOutputFilename = `${camelcase(name)}.test.${
-      typescript ? 'ts' : 'js'
-    }`
-
-    const testFile = await templateForComponentFile({
-      name,
-      extension: typescript ? '.test.ts' : '.test.js',
-      generator: 'directive',
-      templatePath: `${type}.directive.test.ts.template`,
-      outputPath: path.join(
-        getPaths().api.directives,
-        camelName,
-        testOutputFilename,
-      ),
-      templateVars: { camelName },
-    })
-    files.push(testFile)
-  }
-
-  // Returns
-  // {
-  //    "path/to/fileA": "<<<template>>>",
-  //    "path/to/fileB": "<<<template>>>",
-  // }
-  return files.reduce(async (accP, [outputPath, content]) => {
-    const acc = await accP
-
-    const template = typescript
-      ? content
-      : await transformTSToJS(outputPath, content)
-
-    return {
-      [outputPath]: template,
-      ...acc,
-    }
-  }, Promise.resolve({}))
-}
-
-const positionalsObj = {
-  name: {
-    description: 'Name of your directive',
-    type: 'string',
-  },
-}
-
-export const { command, description, builder } =
-  createYargsForComponentGeneration({
-    componentName: 'directive',
-    filesFn: files,
-    positionalsObj,
-    optionsObj: {
-      ...yargsDefaults,
-      type: {
-        type: 'string',
-        choices: ['validator', 'transformer'],
-        description: 'Whether to generate a validator or transformer directive',
-      },
+createYargsForComponentGeneration({
+  componentName: 'directive',
+  optionsObj: {
+    ...yargsDefaults,
+    type: {
+      type: 'string',
+      choices: ['validator', 'transformer'],
+      description: 'Whether to generate a validator or transformer directive',
     },
-  })
+  },
+})
 
-export const handler = async (args) => {
-  recordTelemetryAttributes({
-    command: 'generate directive',
-    type: args.type,
-    force: args.force,
-    rollback: args.rollback,
-  })
+export async function handler(argv) {
+  const { handler: directiveHandler } = await import('./directiveHandler.js')
 
-  let notes = ''
-  const POST_RUN_INSTRUCTIONS = `
-   ${c.note('After modifying your directive, you can add it to your SDLs e.g.:')}
-
-    ${c.info('// example todo.sdl.js')}
-    ${c.info('# Option A: Add it to a field')}
-    type Todo {
-      id: Int!
-      body: String! ${c.tip(`@${args.name}`)}
-    }
-
-    ${c.info('# Option B: Add it to query/mutation')}
-    type Query {
-      todos: [Todo] ${c.tip(`@${args.name}`)}
-    }
-`
-
-  validateName(args.name)
-
-  let directiveType = args.type
-
-  // Prompt to select what type if not specified
-  if (!directiveType) {
-    const response = await prompts({
-      type: 'select',
-      name: 'directiveType',
-      choices: [
-        {
-          value: 'validator',
-          title: 'Validator',
-          description:
-            'Implement a validation: throw an error if criteria not met to stop execution',
-        },
-        {
-          value: 'transformer',
-          title: 'Transformer',
-          description: 'Modify values of fields or query responses',
-        },
-      ],
-      message: 'What type of directive would you like to generate?',
-    })
-
-    directiveType = response.directiveType
-  }
-
-  const tasks = new Listr(
-    [
-      {
-        title: 'Generating directive file ...',
-        task: async () => {
-          const f = await files({ ...args, type: directiveType })
-          return writeFilesTask(f, {
-            overwriteExisting: args.force,
-          })
-        },
-      },
-      {
-        title: 'Generating TypeScript definitions and GraphQL schemas ...',
-        task: () => {
-          // Regenerate again at the end if we rollback changes
-          addFunctionToRollback(async () => {
-            await execa('yarn rw-gen', [], {
-              stdio: 'pipe',
-              shell: true,
-            })
-          }, true)
-          return execa('yarn rw-gen', [], {
-            stdio: 'inherit',
-            shell: true,
-          })
-        },
-      },
-      {
-        title: 'Next steps...',
-        task: () => {
-          // Can't do this, since it strips formatting
-          // task.title = POST_RUN_INSTRUCTIONS
-          // Instead we just console.log the instructions at the end
-          notes = POST_RUN_INSTRUCTIONS
-        },
-      },
-    ].filter(Boolean),
-    { rendererOptions: { collapseSubtasks: false } },
-  )
-
-  try {
-    if (args.rollback && !args.force) {
-      prepareForRollback(tasks)
-    }
-    await tasks.run()
-    if (notes) {
-      console.log(notes)
-    }
-  } catch (e) {
-    console.log(c.error(e.message))
-    process.exit(1)
-  }
+  return directiveHandler(argv)
 }
